refactor: change folders' structure
