+++ src/mocks/index.ts
@@ -1,0 +1,2 @@
+export {default as offersMocks} from './offers';
+export {default as reviewsMocks} from './reviews';
